refactor: POSType
